# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "东南大学成贤学院学生成绩录入及学生进行查询系统的标准API接口文档",
        "title": "成绩录入及查询系统标准接口文档",
        "termsOfService": "暂缺",
        "contact": {
            "name": "黄浩轩",
            "url": "https://github.com/huanghaoxuan",
            "email": "huanghaoxuan1998@outlook.com"
        },
        "license": {
          "name": "GNU GENERAL PUBLIC LICENSE Version 3, 29 June 2007",
          "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:5201",
    "basePath": "/api",
    "paths": {
...
      "/course/delete/{id}": {
        "get": {
          "description": "删除一条课程信息",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "学籍信息"
          ],
          "summary": "删除一条课程信息",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "主键",
              "name": "id",
              "in": "path",
              "required": true
            }
          ]
        }
      },
      "/course/insert": {
        "post": {
          "description": "增加课程信息记录",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "课程信息"
          ],
          "summary": "增加课程信息记录",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "课程名",
              "name": "name",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "学年",
              "name": "year",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "学期",
              "name": "semester",
              "in": "formData"
            }
          ],
          "responses": {
            "200": {
              "description": "{\"status\": 1}",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "/course/selectByPage": {
        "post": {
          "description": "分页查询课程信息记录",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "课程信息"
          ],
          "summary": "分页查询课程信息记录",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "查询页码",
              "name": "pageNum",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "每页条数",
              "name": "pageSize",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "课程名",
              "name": "name",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "学年",
              "name": "year",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "学期",
              "name": "semester",
              "in": "formData"
            }
          ]
        }
      },
      "/course/update": {
        "post": {
          "description": "更新课程信息记录",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "课程信息"
          ],
          "summary": "更新课程信息记录",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "课程名",
              "name": "name",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "学年",
              "name": "year",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "学期",
              "name": "semester",
              "in": "formData"
            }
          ],
          "responses": {
            "200": {
              "description": "{\"status\": 1}",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "/sourceStage/delete/{id}": {
        "get": {
          "description": "删除阶段性测验成绩",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "阶段性测验成绩"
          ],
          "summary": "删除阶段性测验成绩",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "主键",
              "name": "id",
              "in": "path",
              "required": true
            }
          ]
        }
      },
      "/sourceStage/insert": {
        "post": {
          "description": "增加阶段性测验成绩",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "阶段性测验成绩"
          ],
          "summary": "增加阶段性测验成绩",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "学生名字",
              "name": "name",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "学生学号",
              "name": "studentId",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "教学班号",
              "name": "teachingClassId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "阶段性测验id",
              "name": "sourceStageId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "成绩注释",
              "name": "scoresNote",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "成绩",
              "name": "scores",
              "in": "formData"
            }
          ],
          "responses": {
            "200": {
              "description": "{\"status\": 1}",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "/sourceStage/selectByPage": {
        "post": {
          "description": "分页查询阶段性测验成绩",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "阶段性测验成绩"
          ],
          "summary": "分页查询阶段性测验成绩",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "查询页码",
              "name": "pageNum",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "每页条数",
              "name": "pageSize",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "学生名字",
              "name": "name",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "学生学号",
              "name": "studentId",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "教学班号",
              "name": "teachingClassId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "阶段性测验id",
              "name": "sourceStageId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "成绩注释",
              "name": "scoresNote",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "成绩",
              "name": "scores",
              "in": "formData"
            }
          ]
        }
      },
      "/sourceStage/update": {
        "post": {
          "description": "更新阶段性测验成绩",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "阶段性测验成绩"
          ],
          "summary": "更新阶段性测验成绩",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "学生名字",
              "name": "name",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "学生学号",
              "name": "studentId",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "教学班号",
              "name": "teachingClassId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "阶段性测验id",
              "name": "sourceStageId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "成绩注释",
              "name": "scoresNote",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "成绩",
              "name": "scores",
              "in": "formData"
            }
          ],
          "responses": {
            "200": {
              "description": "{\"status\": 1}",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "/sourceStage/updates": {
        "post": {
          "description": "批量更新阶段性测验成绩",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "阶段性测验成绩"
          ],
          "summary": "批量更新阶段性测验成绩",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "学生名字",
              "name": "name",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "学生学号",
              "name": "studentId",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "教学班号",
              "name": "teachingClassId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "阶段性测验id",
              "name": "sourceStageId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "成绩注释",
              "name": "scoresNote",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "成绩",
              "name": "scores",
              "in": "formData"
            }
          ],
          "responses": {
            "200": {
              "description": "{\"status\": 1}",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
}